Rename: 메인 페이지 관련 폴더 구조 수정
