+++ src/pages/index.ts
@@ -3,3 +3,4 @@ export * from './group-management'
 export * from './point'
 export * from './auth'
 export * from './profile'
+export * from './main'
