+++ src/pages/main/components/game-section/CommonGameSection.tsx
@@ -4,24 +4,24 @@ import { useQuery } from '@tanstack/react-query'
 import { friendsQueries } from '@/api/services/friend/queries'
 import { Loading } from '@/components/Loading'
 
-import { MainSection } from './MainSection'
+import { CommonMain } from './common-main'
 
-interface CommonGameProps {
+interface CommonGameSectionProps {
   handleFinishGame: () => void
   handleClickProfile: () => void
 }
 
-export const CommonGame = ({
+export const CommonGameSection = ({
   handleFinishGame,
   handleClickProfile,
-}: CommonGameProps) => {
+}: CommonGameSectionProps) => {
   const { data: friends, status } = useQuery(friendsQueries.myFriends())
 
   if (status === 'pending') return <Loading />
   if (!friends) return <Text>친구를 초대해보세요!</Text>
 
   return (
-    <MainSection
+    <CommonMain
       friends={friends}
       onFinsihGame={handleFinishGame}
       onClickProfile={handleClickProfile}
