+++ src/pages/main/components/game-section/common-main/index.ts
@@ -1,0 +1,1 @@
+export { CommonMain } from './CommonMain'
