+++ src/pages/main/components/game-section/group-main/GroupMain.tsx
@@ -1,59 +1,61 @@
 import { useEffect, useState } from 'react'
 import { BiChevronsRight, BiGroup } from 'react-icons/bi'
+import { Link } from 'react-router-dom'
 
 import { Button, Flex, Heading, Text } from '@chakra-ui/react'
 import { useMutation } from '@tanstack/react-query'
 
 import { queryClient } from '@/api/instance'
 import {
-  AnswerQuestionParam,
-  answerRandomQuestion,
+  AnswerGroupQuestionParams,
+  answerGroupQuestion,
 } from '@/api/services/answer/question.api'
-import { useRandomQuestion } from '@/api/services/question/random.api'
+import { useGroupRandomQuestion } from '@/api/services/question/random.api'
 import { pointQuries } from '@/api/services/user/point.api'
 import { Loading } from '@/components/Loading'
 import { useProfileRandom } from '@/hooks/useProfileRandom'
-import { Friend } from '@/types'
+import { Member } from '@/types'
 
-import ProfileGrid from './ProfileGrid'
+import { ProfileGrid } from './profile-grid'
 
-interface MainSectionProps {
-  friends: Friend[]
+interface GroupMainProps {
+  groupId: number
+  members: Member[]
   onFinsihGame: () => void
   onClickProfile: () => void
 }
 
 const QUESTION_SIZE = 5
 
-export const MainSection = ({
-  friends,
+export const GroupMain = ({
+  groupId,
+  members,
   onFinsihGame,
   onClickProfile,
-}: MainSectionProps) => {
+}: GroupMainProps) => {
   const {
     data: questions,
     status,
     refetch,
-  } = useRandomQuestion({
-    size: QUESTION_SIZE,
-  })
+  } = useGroupRandomQuestion({ groupId })
   const { mutate: answerQuestion } = useMutation({
-    mutationFn: (params: AnswerQuestionParam) => answerRandomQuestion(params),
+    mutationFn: (params: AnswerGroupQuestionParams) =>
+      answerGroupQuestion(params),
     onSuccess: () => {
       queryClient.invalidateQueries({ queryKey: pointQuries.all() })
     },
   })
-  const { pickedProfiles, reloadRandomProfiles } = useProfileRandom(friends)
+  const { pickedProfiles, reloadRandomProfiles } = useProfileRandom(members)
 
   const [questionSize, setQuestionSize] = useState(QUESTION_SIZE)
   const [questionIndex, setQuestionIndex] = useState(0)
 
   useEffect(() => {
-    if (questionIndex === QUESTION_SIZE) {
+    if (questionIndex === questionSize) {
       onFinsihGame()
       refetch()
     }
-  }, [questionIndex, onFinsihGame, refetch])
+  }, [questionIndex, onFinsihGame, refetch, questionSize])
 
   useEffect(() => {
     if (questions?.length && questions.length < QUESTION_SIZE) {
@@ -71,12 +73,15 @@ export const MainSection = ({
         textAlign="center"
         gap={5}
       >
-        <Heading size="lg">질문이 없습니다😢</Heading>
-        <Heading size="md">관리자에게 문의해주세요</Heading>
+        <Heading size="lg">그룹 질문이 없습니다😢</Heading>
+        <Heading size="md">그룹 페이지에서 질문을 추가해보세요!</Heading>
+        <Link to={`/group/${groupId}`}>
+          <Button colorScheme="primary" width="full" height="2.5rem">
+            그룹 페이지로 이동하기
+          </Button>
+        </Link>
       </Flex>
     )
-
-  if (questionIndex === questionSize) return <Loading />
 
   const handleQuestionSkip = () => {
     reloadRandomProfiles()
@@ -86,17 +91,18 @@ export const MainSection = ({
   const handleProfileSelect = (pickedId: number) => {
     handleQuestionSkip()
     answerQuestion({
+      groupId,
       questionId: questions[questionIndex].questionId,
       pickedId,
     })
     onClickProfile()
   }
 
-  if (questionIndex === QUESTION_SIZE) return <Loading />
+  if (questionIndex === questionSize) return <Loading />
 
   return (
     <Flex
-      height="full"
+      height="100%"
       flexDirection="column"
       justifyContent="space-between"
       textAlign="center"
