+++ src/pages/main/components/game-section/group-main/index.ts
@@ -1,0 +1,1 @@
+export { GroupMain } from './GroupMain'
