+++ src/pages/main/components/game-section/group-main/profile-grid/index.ts
@@ -1,0 +1,1 @@
+export { ProfileGrid } from './ProfileGrid'
