+++ src/pages/main/components/game-section/index.ts
@@ -1,0 +1,2 @@
+export { GroupGameSection } from './GroupGameSection'
+export { CommonGameSection } from './CommonGameSection'
