+++ src/pages/main/components/index.ts
@@ -1,0 +1,3 @@
+export * from './game-section'
+export * from './ready-section'
+export * from './replay-section'
