+++ src/pages/main/components/ready-section/ReadySection.tsx
