+++ src/pages/main/components/ready-section/index.ts
@@ -1,0 +1,1 @@
+export { ReadySection } from './ReadySection'
