+++ src/pages/main/components/replay-section/index.ts
@@ -1,0 +1,1 @@
+export { ReplaySection } from './ReplaySection'
